docs: fix threshold parameter types for NetworkBandwidth
threshold_send and threshold_receive were declared as ParameterType.INTEGER
but are read with config.getfloat(), accepting fractional byte rates.
Change the declared type to FLOAT to match the actual runtime behaviour.

+++ src/autosuspend/checks/linux.py
@@ -142,13 +142,13 @@ class Load(Activity):
 )
 @config_param(
     "threshold_send",
-    ParameterType.INTEGER,
+    ParameterType.FLOAT,
     "If the average sending bandwidth of one of the specified interfaces is above this threshold, then activity is detected. Specified in bytes/s",
     default=100,
 )
 @config_param(
     "threshold_receive",
-    ParameterType.INTEGER,
+    ParameterType.FLOAT,
     "If the average receive bandwidth of one of the specified interfaces is above this threshold, then activity is detected. Specified in bytes/s",
     default=100,
 )
